requirements.txt
python-dotenv==0.19.0
Flask==1.1.2
Flask-RESTful==0.3.8
requests==2.26.0
gunicorn==20.1.0
influxdb-client==1.21.0
